# ollantaweb/docs/swagger.yaml
basePath: /
definitions:
...
  api.groupListResponse:
    properties:
      groups:
        items:
          $ref: '#/definitions/api.groupView'
        type: array
    type: object
  api.groupMembersResponse:
    properties:
      members:
        items:
          $ref: '#/definitions/api.userView'
        type: array
    type: object
  api.groupView:
    properties:
      description:
        type: string
      id:
        type: integer
      is_builtin:
        type: boolean
      name:
        type: string
    type: object
...
  api.issueTrackingBackfillResult:
    properties:
      issues_updated:
        type: integer
      project_key:
        type: string
      scans_processed:
        type: integer
      scopes_processed:
        type: integer
      unknown_issues:
        type: integer
    type: object
...
  api.userView:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      login:
        type: string
      name:
        type: string
      provider:
        type: string
    type: object
...
info:
  contact: {}
  description: Ollanta REST API documentation.
  title: Ollanta API
  version: "1.0"
paths:
...
  /api/v1/groups:
    get:
      description: Returns all groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.groupListResponse'
      summary: List groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Create a new group
      parameters:
      - description: Group data
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.groupView'
      summary: Create group
      tags:
      - groups
  /api/v1/groups/{id}:
    delete:
      description: Delete a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete group
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Update a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group data
        in: body
        name: body
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.groupView'
      summary: Update group
      tags:
      - groups
  /api/v1/groups/{id}/members:
    get:
      description: List members of a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.groupMembersResponse'
      summary: List group members
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Add a user to a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Member data
        in: body
        name: body
        required: true
        schema:
          properties:
            user_id:
              format: int64
              type: integer
          type: object
      responses:
        "204":
          description: No Content
      summary: Add group member
      tags:
      - groups
  /api/v1/groups/{id}/members/{uid}:
    delete:
      description: Remove a user from a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: path
        name: uid
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Remove group member
      tags:
      - groups
...
  /api/v1/projects/{key}/issues/backfill-tracking-state:
    post:
      description: Backfill issue tracking states for a project
      parameters:
      - description: Project key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.issueTrackingBackfillResult'
      summary: Backfill tracking state
      tags:
      - issues
...
swagger: "2.0"